deleted game canvas and moved its content to GamePlayer component
The reason why i moved it here is because i need to access the gameref inside of the Game player so that i can link the cabinet buttons to the game state so that the cabinet buttons could affect the game state
